+++ example/generate-open-api-schema.ts
@@ -1,9 +1,9 @@
-import {generateOpenApi} from '../src';
+import {OpenAPI} from '../src';
 import {routing} from './routing';
 import {version} from '../package.json';
 
-console.log(generateOpenApi({
+console.log(new OpenAPI({
   routing, version,
   title: 'Example API',
   serverUrl: 'http://example.com'
-}).getSpecAsYaml());
+}).builder.getSpecAsYaml());