+++ src/index.ts
@@ -7,9 +7,9 @@ export { createMiddleware } from './middleware';
 export { ResultHandler } from './result-handler';
 export { Routing, initRouting } from './routing';
 export { createServer } from './server';
-export { generateOpenApi } from './open-api';
+export { OpenAPI } from './open-api';
 
-import * as z from 'zod';
+import { z } from 'zod';
 import * as createHttpError from 'http-errors';
 
 export { createHttpError, z };