+++ tests/unit/open-api.spec.ts
@@ -1,15 +1,45 @@
 import {routing} from '../../example/routing';
-import {generateOpenApi} from '../../src';
+import {endpointsFactory} from '../../example/factories';
+import {z, OpenAPI} from '../../src';
 
 describe('Open API generator', () => {
   describe('generateOpenApi()', () => {
     test('should generate the correct schema of example routing', () => {
-      const spec = generateOpenApi({
+      const spec = new OpenAPI({
         routing,
         version: '1.2.3',
         title: 'Example API',
         serverUrl: 'http://example.com'
-      }).getSpecAsYaml();
+      }).builder.getSpecAsYaml();
+      expect(spec).toMatchSnapshot();
+    });
+
+    test('should generate the correct schema for complex types', () => {
+      const literalValue = 'something';
+      const spec = new OpenAPI({
+        routing: {
+          v1: {
+            getSomething: endpointsFactory.build({
+              methods: ['get'],
+              input: z.object({
+                array: z.array(z.number().int().positive()),
+                transformer: z.string().transform((str) => str.length)
+              }),
+              output: z.object({
+                literal: z.literal(literalValue),
+                transformation: z.number(),
+              }),
+              handler: async ({input}) => ({
+                literal: literalValue as typeof literalValue,
+                transformation: input.transformer,
+              })
+            })
+          }
+        },
+        version: '3.4.5',
+        title: 'Testing Complex Types',
+        serverUrl: 'http://example.com'
+      }).builder.getSpecAsYaml();
       expect(spec).toMatchSnapshot();
     });
   });